disaster_backend/requirements.txt
asgiref==3.9.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.7.14
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
git-filter-repo==2.47.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.5.1
numpy==2.3.2
pandas==2.3.1
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
scikit-learn==1.7.1
scipy==1.16.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==12.4.0
threadpoolctl==3.6.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
zope.interface==7.2
